feat: Añadir configuración de ESLint para deshabilitar la regla de react-refresh en componentes de UI y reestructurar la tabla de cuotas en componentes separados
